+++ src/main/java/me/salamander/noisetest/gui/Parameter.java
@@ -1,4 +1,4 @@
 package me.salamander.noisetest.gui;
 
-public record Parameter(String name, int index, double minValue, double maxValue, double step) {
+public record Parameter(String name, int index, float minValue, float maxValue, float step) {
 }
