+++ src/main/java/me/salamander/noisetest/modules/NoiseModule.java
@@ -1,7 +1,7 @@
 package me.salamander.noisetest.modules;
 
 public interface NoiseModule {
-    double sample(double x, double y); //TODO: Make this take and return floats. Double is not needed
+    float sample(float x, float y); //TODO: Make this take and return floats. Double is not needed
     default void setSeed(long s){
 
     }
